**v.0.3.6**
* bug fix `assignment_ngs` during imputation, the imputation module could not recognise that REF/ALT allele
are not necessary or usefull for assignment analysis.
*enhancement to `assignment_ngs` and `assignment_mixture` so that when
`marker.number` include `"all"` the `iteration.method` is set automatically to `1`
when conducting the assignment with all the markers.
Iterations at this point is useless.
* random seed number is now stored in the appropriate files.
* `assignment_mixture`: with `assignment.analysis = "gsi_sim` the unknown/mixture samples
are compared with baseline populations using common markers between the pair.
Now, the tables include the number of markers used. The summary provides the mean
number of markers. This number will change each time randomness is used.